+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Tous les changements apportés au projet sont suivis dans ce fichier.
 Le format est basé sur [Tenez un Changelog](https://keepachangelog.com/fr/1.0.0/).
 
 ## [À venir dans la prochaine version]
+# Réparé
+- Répare bug quand on modifie un temps passé sur un projet dont on n'est plus contributeur